requirement.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain_community
bs4
scikit-learn
nltk
-e.